+++ CHANGELOG.md
@@ -2,6 +2,7 @@
   <ul>
     <li><code>BigDecimal|BigInteger</code> constant inspection: full number parsing (e.g. "0" = "0000.0000" = "0e10") and Java 9's <code>BigInteger.TWO</code> constant support</li>
     <li>Removed “Nullable argument to string concatenation” because of false-positives</li>
+    <li>Stopped reporting <code>AnyEnum.values()</code> calls in non-static context, leave this to Kotlin team and their new <code>AnyEnum.entries</code> feature</li>
   </ul>
 
 ### 0.28