+++ CHANGELOG.md
@@ -6,6 +6,7 @@
     <li>Reporting <code>new OkHttpClient()</code> in “Allocation should be cached” inspection (and not reporting <code>Gson.Builder()</code>, it could be configured somehow)</li>
     <li>Previewing even less opaque colors, >= 0x19 instead of > 0x27, useful for Ripple</li>
     <li>Don't showing useless interface hints for anonymous classes</li>
+    <li>Reporting heavyweight property delegation for <code>var</code>s, not only <code>val</code>s</li>
   </ul>
 
 ### 0.28