+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+
+### 0.30
+  <ul>
+  </ul>
+
 ### 0.29
   <ul>
     <li><code>BigDecimal|BigInteger</code> constant inspection: full number parsing (e.g. "0" = "0000.0000" = "0e10") and Java 9's <code>BigInteger.TWO</code> constant support</li>